pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "open-agent-id"
version = "0.2.0"
description = "Python SDK for Open Agent ID — register, sign, and verify AI agent identities"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
keywords = ["agent", "identity", "did", "ed25519", "verification", "ai-agent", "open-agent-id"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security :: Cryptography",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "httpx>=0.27",
    "pynacl>=1.5",
    "cryptography>=42.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.urls]
Homepage = "https://openagentid.org"
Repository = "https://github.com/openagentid/agent-id-python"

[tool.hatch.build.targets.wheel]
packages = ["agent_id"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
